COMP: Support setting default library cxx wrapper outside of SlicerExecutionModelConfig.cmake
Currently, it is possible to configure the SlicerExecutionModel so that
default value configure in the SlicerExecutionModelConfig.cmake file are used.

This is done by configuring the project passing option of the form:
  SlicerExecutionModel_DEFAULT_CLI_LIBRARY_WRAPPER_CXX

This commit extend the idea by allowing to also specify "default" value
independently of so the default value configured in the cache are not
considered. The following variable can be set:

  SlicerExecutionModel_CLI_LIBRARY_WRAPPER_CXX