COMP: Update GenerateCLP and ModuleDescriptionParser to use ITKv4
This commit explicitly specifies the list of components to link with. For
convenience, it also exports (in the config file) a variable
ModuleDescriptionParser_ITK_COMPONENTS allowing dependent library in other
project to find ITK specifying these components and include them in the
list of target libraries.

Additionally, since ModuleDescriptionParser does NOT have a proper
config file (this will be fixed at a later time), the GenerateCLP USE_FILE
is updated to import the targets expected by ModuleDescriptionParser.